--- test/reports/xml_result.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<testsuites>
-	<testsuite name="TestXML-20220310153856" tests="4" file=".py" time="0.006" timestamp="2022-03-10T15:38:56" failures="1" errors="1" skipped="1">
-		<testcase classname="TestXML" name="test_success" time="0.001" timestamp="2022-03-10T15:38:56" file="test_xml_report.py" line="9">
-			<!--执行成功-->
-		</testcase>
-		<testcase classname="TestXML" name="test_fail" time="0.000" timestamp="2022-03-10T15:38:56" file="test_xml_report.py" line="17">
-			<failure type="AssertionError" message="5 != 6"><![CDATA[Traceback (most recent call last):
-  File "test_xml_report.py", line 18, in test_fail
-    self.assertEqual(5, 6)
-AssertionError: 5 != 6
-]]></failure>
-		</testcase>
-		<testcase classname="TestXML" name="test_error" time="0.005" timestamp="2022-03-10T15:38:56" file="test_xml_report.py" line="20">
-			<error type="NameError" message="name 'a' is not defined"><![CDATA[Traceback (most recent call last):
-  File "test_xml_report.py", line 21, in test_error
-    self.assertEqual(a, 6)
-NameError: name 'a' is not defined
-]]></error>
-		</testcase>
-		<testcase classname="TestXML" name="test_skip" time="0.000" timestamp="2022-03-10T15:38:56" file="test_xml_report.py" line="13">
-			<skipped type="skip" message="skip case"/>
-		</testcase>
-	</testsuite>
-	<testsuite name="TestXML2-20220310153856" tests="1" file=".py" time="0.000" timestamp="2022-03-10T15:38:56" failures="0" errors="0" skipped="0">
-		<testcase classname="TestXML2" name="test_success" time="0.000" timestamp="2022-03-10T15:38:56" file="test_xml_report.py" line="26"/>
-	</testsuite>
-	<testsuite name="TestXML3-20220310153856" tests="1" file=".py" time="0.000" timestamp="2022-03-10T15:38:56" failures="1" errors="0" skipped="0">
-		<testcase classname="TestXML3" name="test_fail" time="0.000" timestamp="2022-03-10T15:38:56" file="test_xml_report.py" line="32">
-			<failure type="AssertionError" message="3 != 4"><![CDATA[Traceback (most recent call last):
-  File "test_xml_report.py", line 33, in test_fail
-    self.assertEqual(3, 4)
-AssertionError: 3 != 4
-]]></failure>
-		</testcase>
-	</testsuite>
-</testsuites>
