+++ tests/reports/test_unit_xml.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<testsuites>
+	<testsuite name="TestDemo-20230306233452" tests="4" file=".py" time="0.005" timestamp="2023-03-06T23:34:52" failures="1" errors="1" skipped="1">
+		<testcase classname="TestDemo" name="test_success" time="0.004" timestamp="2023-03-06T23:34:52" file="test_unit_xml.py" line="8">
+			<doc><![CDATA[执行成功]]></doc>
+		</testcase>
+		<testcase classname="TestDemo" name="test_fail" time="0.000" timestamp="2023-03-06T23:34:52" file="test_unit_xml.py" line="16">
+			<failure type="AssertionError" message="5 != 6"><![CDATA[Traceback (most recent call last):
+  File ".\test_unit_xml.py", line 17, in test_fail
+    self.assertEqual(5, 6)
+AssertionError: 5 != 6
+]]></failure>
+		</testcase>
+		<testcase classname="TestDemo" name="test_error" time="0.000" timestamp="2023-03-06T23:34:52" file="test_unit_xml.py" line="19">
+			<error type="NameError" message="name 'a' is not defined"><![CDATA[Traceback (most recent call last):
+  File ".\test_unit_xml.py", line 20, in test_error
+    self.assertEqual(a, 6)
+NameError: name 'a' is not defined
+]]></error>
+		</testcase>
+		<testcase classname="TestDemo" name="test_skip" time="0.001" timestamp="2023-03-06T23:34:52" file="test_unit_xml.py" line="12">
+			<skipped type="skip" message="skip case"/>
+		</testcase>
+	</testsuite>
+</testsuites>
